(rename) : 패키지명 변경 #EAT-118
