+++ src/main/java/team/gwon/haveameal/member/registrationservice/personaldataencryption/SecurityConfig.java
@@ -1,4 +1,4 @@
-package team.gwon.haveameal.member.encryptionservice.personaldata;
+package team.gwon.haveameal.member.registrationservice.personaldataencryption;
 
 import java.security.SecureRandom;
